componetized Notes and added some styles
