+++ notes-app/src/components/Notes/index.jsx
@@ -1,0 +1,30 @@
+import { useDispatch, useSelector } from "react-redux";
+import { toggleImportanceOf } from "../../reducers/noteReducer";
+
+export const Notes = () => {
+  const notes = useSelector((state) => state);
+  const dispatch = useDispatch();
+
+  const toggleImportant = (id) => {
+    dispatch(toggleImportanceOf(id));
+  };
+
+  return (
+    <div>
+      <ul>
+        {notes.map((note) => {
+          return (
+            <li
+              style={{ cursor: "pointer" }}
+              key={note.id}
+              onClick={() => toggleImportant(note.id)}
+            >
+              <p>{note.content}</p>
+              <strong>{note.important ? "important" : "not important"}</strong>
+            </li>
+          );
+        })}
+      </ul>
+    </div>
+  );
+};
