+++ notes-app/src/index.js
@@ -4,6 +4,7 @@ import { App } from "./App";
 import { Provider } from "react-redux";
 import { noteReducer } from "./reducers/noteReducer";
 import { configureStore } from "@reduxjs/toolkit";
+import styles from "./styles/global.css";
 
 const store = configureStore({ reducer: noteReducer });
